update slide HomePage and add map
